Module delanocreds::attributes

A module for creating:

Individual Attibute, a Multihash generated from bytes (likely from a string)

Entry, a vector of Attibutes

AttributeEntries (a Vector of attribute entries) hierarchy starts with Root Issuer entry/entries

AttributeEntries can be extended up to k_prime if update_key is provided

AttributeEntries can be restricted by zeroizing opening information

AttributeEntries can hold up to a cumulative total of max_cardinality attributes

There can be up to message_l entries in each AttributeEntries vector

Each entry in AttributeEntries is a vector holding a number of Attributes

You can select subsets of AttributesVectors, by indicating which indexes you want to select, a 2D vector matix

When opening information is zeroized, then the corresponding entry in AttributeEntries cannot be selected, thus opening and atrributes have a relationship with each other. The opening vector is held in the credential, ths the credential and the attributes have a relationship with each other.

API

// create a new Attribute let attribute: Attribute = attribute(“age>21”)

// attributes have a digest() method which returns a &u8 attributes.push(vec![“a”.to_string(), “b”.to_string(), “c”.to_string()]);

// select from the attributes let selected_attrs = attributes.select(vec![vec![], vec![0, 1], vec![0, 1]]);

Structs

Attribute

Functions

attribute

Converts any type that can be converted to a byte slice into a Content Identifier (CID) using the SHAKE256 hash function with length 48.

from_cid

Generate an attribute from a CID Check to verify that the CID is a SHAKE256 hash with length 48