Module delanocreds::config

Configuration of the Root Issuer and the Credential Issuer.

Root Issuer

The Root Issuer is the entity that issues the root credential. It is the entity that is trusted by the verifier to issue the root credential. The root credential is used to issue the credential schema and the credential definition.

Root Issuer Configuration

To make a root issuer, you need to create a RootIssuerConfig struct. This struct contains the following fields:

name: The name of the root issuer.

Credential Issuer

Constants

CHALLENGE_STATE_NAME

DEFAULT_MAX_CARDINALITY

Default Max Cardinality: The maximum number of total attribute elements allowed in a credential.

DEFAULT_MAX_ENTRIES

Default Max Attributes: The maximum number of attribute entries allowed in a credential.