Module delanocreds::keypair

This is implementation of delegatable anonymous credential using SPSQE-UC signatures and set commitment. See the following for the details:

Practical Delegatable Anonymous Credentials From Equivalence Class Signatures, PETS 2023. https://eprint.iacr.org/2022/680

Modules

spseq_uc

Structs

CredOffer

CredOffer (cred_R)

CredProof

Credentials Proof

DelegatedCred

DelegateeCred

MaxCardinality

MaxEntries

Nym

Nym

NymProof

NymProof

NymPublic

Pseudonym Public information

Signature

Signer

Root Issuer aka Certificate Authority

UserKey

Enums

VK

Verification Key This key has elements from both G1 and G2, so to make a Vector of VK, we need to use enum

Functions

verify

Verifies a signature for a given message set

verify_proof

Verify proof of a credential