Module delanocreds::keypair::spseq_uc

Structs

Credential

A Credential is an EqcSignature signature returned by the sign function It contains the sigma, update key, commitment vector

RandomizedPubKey

Enums

UpdateError

Functions

change_rel

Change Relations of a Credential. Push additional Entry onto the end of Message Commitment Stack.

change_rep

Change the Representative of the signature message pair to a new commitment vector and user public key. This is used to update the signature message pair to a new user public key. The new commitment vector is computed using the old commitment vector and the new user public key. The new user public key is computed using the old user public key and the update key. The update key is computed during the signing process.

rndmz_pk

Type Definitions

OpeningInformation

UpdateKey

Update Key alias