Function delanocreds::keypair::spseq_uc::change_rep

pub fn change_rep(
    pk_u: &G1,
    cred: &Credential,
    mu: &FieldElement,
    psi: &FieldElement,
    extendable: bool
) -> (RandomizedPubKey, Credential, FieldElement)

Change the Representative of the signature message pair to a new commitment vector and user public key. This is used to update the signature message pair to a new user public key. The new commitment vector is computed using the old commitment vector and the new user public key. The new user public key is computed using the old user public key and the update key. The update key is computed during the signing process.

Arguments

pk_u: Current public key associated to the Credential’s Sigma Signature

cred: Credential to be updated

mu: randomness is used to randomize commitment vector and signature accordingly. Should be set to one (1) unless it’s the final use, as it cannot be re-randomized again.

psi: randomness is used to randomize commitment vector and signature accordingly

extendable: a flag to determine if it needs to refresh (randomize) the update_key as well or not. Only takes effect if there is both b and an orig_sig.update_key

Returns

returns an updated signature σ for a new commitment vector and corresponding openings