Struct delanocreds::keypair::Nym

pub struct Nym {
    pub public: NymPublic,
    /* private fields */
}

Nym

secret: Secret, secret witness of the user

pub public: NymPublic, proof of the user

Fields
public: NymPublic
Implementations

impl Nym

pub fn new( nym_public_key: RandomizedPubKey, secret_wit: Secret<FieldElement>, public_parameters: ParamSetCommitment ) -> Nym

Create a new Nym associated to the Public Parameters (ParamSetCommitment) of the Issuer

Arguments

nym_public_key: RandomizedPubKey, public key of the Nym

secret_wit: Secret<[FieldElement]>, secret witness of the user

public_parameters: ParamSetCommitment, public parameters of the user

pub fn offer_builder<'a>( &'a self, cred: &'a Credential, entries: &[Entry] ) -> OfferBuilder<'_>

Creates an super::OfferBuilder for a Credential using this Nym

Arguments

cred: Credential, credential being offered to the delegatee

entries: &Entry, vector of all attribute Entrys corresponding to the cred

Returns

super::OfferBuilder

pub fn proof_builder<'a>( &'a self, cred: &'a Credential, entries: &'a [Entry] ) -> ProofBuilder<'_>

Create a super::ProofBuilder for a Credential and Entrys using this Nym

Arguments

cred: Credential, credential being proven

entries: &Entry, vector of all attribute Entrys corresponding to the cred

pub fn offer( &self, cred: &Credential, addl_attrs: &Option<Entry>, their_nym: &NymPublic ) -> Result<Offer, Error>

Creates an orphan Offer for a Credential.

Important Note: The orphan Signature created by this Offer can be used by any holder who also has possesion of the associated crate::attributes::Attributes, so it is important to either keep Offers confidential or disassociated with the crate::attributes::Attributes. In other words, do not make both pubically available and associated; do not store them in the same place or object, unless it is with the intended Holder.

pub fn accept(&self, offer: &Offer) -> Credential

Accept the credential and adapt Signature back to Self.

Delegated User (delegatee) uses this function to assign to their pseudonym.

Arguments

offer: Offer, credential received from delegator

Returns

Credential with the new Signature for the our Nym’s public key

pub fn prove( &self, cred: &Credential, all_attributes: &[Entry], selected_attrs: &[Entry] ) -> CredProof

Prove selected crate::attributes::Attributes for a Credential, given all Entrys

If a selected Entry is empty, it is skipped (not proven) and there is no need to provide any of the corresponding full/complete Entry as prove() filters out any selected Entrys which are empty and excludes them from the proof.

Arguments

cred: &Credential, credential of the user

all_attributes: &Entry, vector of complete attribute Entrys of the user corresponding to the selected attributes

selected_attrs: &Entry, vector of selected attribute Entrys to be disclosed