Struct delanocreds::set_commits::CrossSetCommitment

pub struct CrossSetCommitment {
    pub param_sc: ParamSetCommitment,
}

CrossSetCommitment extends the Set Commitment to provide aggregation witness and a batch verification

param_sc: ParamSetCommitment public parameters for the commitment scheme, as P^ai, P_hat^ai, P = g1, P_hat = g2, Order, BG

Fields
param_sc: ParamSetCommitment
Implementations

impl CrossSetCommitment

CrossSetCommitment is a commitment scheme that extends the Set Commitment to provide aggregation witness and a batch verification

pub fn aggregate_cross(witness_vector: &[G1], commit_vector: &[G1]) -> G1

Computes an aggregate proof of valid subsets of a set of messages.

Arguments

witness_vector: a vector of witnessess commit_vector: the commitment vector

Returns

a proof which is a aggregate of witnesses and shows all subsets are valid for respective sets

pub fn verify_cross( param_sc: &ParamSetCommitment, commit_vector: &[G1], entry_subsets_vector: &[Entry], proof: &G1 ) -> Result<bool, SerzDeserzError>

Verifies an aggregate proof of valid subsets of a set of messages.

Arguments

param_sc: public parameters commit_vector: the commitment vector entry_subsets_vector: Vector of sleected Entrys proof: a proof which is a aggregate of witnesses

Returns

true if the proof is valid, false otherwise