Trait delanocreds::set_commits::Commitment

pub trait Commitment {
    // Required methods
    fn new(t: MaxCardinality) -> Self;
    fn public_parameters(self) -> ParamSetCommitment;

    // Provided methods
    fn setup(max_cardinality: MaxCardinality) -> ParamSetCommitment { ... }
    fn commit_set(
        param_sc: &ParamSetCommitment,
        mess_set_str: &Entry
    ) -> Result<(G1, FieldElement), SerzDeserzError> { ... }
    fn open_set(
        param_sc: &ParamSetCommitment,
        commitment: &G1,
        open_info: &FieldElement,
        mess_set_str: &Entry
    ) -> Result<bool, SerzDeserzError> { ... }
    fn open_subset(
        param_sc: &ParamSetCommitment,
        all_messages: &Entry,
        open_info: &FieldElement,
        subset: &Entry
    ) -> Result<Option<G1>, SerzDeserzError> { ... }
    fn verify_subset(
        param_sc: &ParamSetCommitment,
        commitment: &G1,
        subset_str: &Entry,
        witness: &G1
    ) -> Result<bool, SerzDeserzError> { ... }
}

Public trait Commitment common for both SetCommitment and CrossSetCommitment This trait is used to define the common functions for both SetCommitment and CrossSetCommitment

Required Methods

fn new(t: MaxCardinality) -> Self

Creates a new commitment scheme.

Arguments

t: max cardinality of the set. Cardinality of the set is in [1, t]. t is a public parameter.

Returns

A Commitment scheme, either SetCommitment or CrossSetCommitment

fn public_parameters(self) -> ParamSetCommitment

Public parameters of the commitment scheme.

Provided Methods

fn setup(max_cardinality: MaxCardinality) -> ParamSetCommitment

Generates a new commitment scheme.

fn commit_set( param_sc: &ParamSetCommitment, mess_set_str: &Entry ) -> Result<(G1, FieldElement), SerzDeserzError>

Commit to a set of messages

Arguments

param_sc: ParamSetCommitment public parameters for the commitment scheme, as P^ai, P_hat^ai, P = g1, P_hat = g2, Order, BG mess_set_str: a vector of Entrys

Returns

Tuple of (commitment, witness) or a SerzDeserzError if

Method

Convert the message set to a vector of FieldElements

Compute the commitment as a product of P^ai, P_hat^ai, where ai is the ith element of the message set

Compute the witness as a product of P^ai, P_hat^ai, where ai is the ith element of the message set

Return the commitment and witness

fn open_set( param_sc: &ParamSetCommitment, commitment: &G1, open_info: &FieldElement, mess_set_str: &Entry ) -> Result<bool, SerzDeserzError>

Open a commitment to a set of messages. This is the verification step. Verifies the opening information of a set.

Arguments

param_sc: public parameters for the commitment scheme, as P^ai, P_hat^ai, P = g1, P_hat = g2, Order, BG commitment: commitment to a set of messages open_info: opening information of the commitment mess_set_str: a message set as a string

Returns

bool: true if the opening information is valid, false otherwise

fn open_subset( param_sc: &ParamSetCommitment, all_messages: &Entry, open_info: &FieldElement, subset: &Entry ) -> Result<Option<G1>, SerzDeserzError>

Opens a subet of Attributes for an Entry open_subset generates a witness for the subset only if the length of the subset is less than the length of the message set

Arguments

param_sc: public parameters for the commitment scheme, as P^ai, P_hat^ai, P = g1, P_hat = g2, Order, BG mess_set_str: the message set open_info: opening information of the commitment subset_str: the subset of the message set

Returns

Some witness for the subset or None if the subset is not a subset of the message set

fn verify_subset( param_sc: &ParamSetCommitment, commitment: &G1, subset_str: &Entry, witness: &G1 ) -> Result<bool, SerzDeserzError>

VerifySubset verifies the witness for the subset. Verifies if witness proves that subset_str is a subset of the original message set.

Arguments

param_sc: public parameters for the commitment scheme, as P^ai, P_hat^ai, P = g1, P_hat = g2, Order, BG commitment: commitment to a set of messages subset_str: the subset of the message set witness: witness for the subset

Returns

bool: true if the witness is valid, false otherwise